Organize the classes into packages
