+++ src/main/java/lawbringer/Deadline.java
@@ -1,3 +1,5 @@
+package lawbringer;
+
 public class Deadline extends Task {
 
     protected String by;
