+++ src/main/java/lawbringer/LawbringerException.java
@@ -1,4 +1,6 @@
-public class LawbringerException extends Exception{
+package lawbringer;
+
+public class LawbringerException extends Exception {
     public LawbringerException(String exception) {
         super(exception);
     }
