+++ src/main/java/lawbringer/Task.java
@@ -1,3 +1,5 @@
+package lawbringer;
+
 public abstract class Task {
 
     protected String title;
